change int to unsigned int

+++ core/shaders/fragment/pathTracer/random.glsl
@@ -1,4 +1,4 @@
-uint Seed = uint((TexCoords.x + TexCoords.y * TexCoords.x) * 549856.0) + uint(FrameCount) * 5458u;
+uint Seed = uint((TexCoords.x + TexCoords.y * TexCoords.x) * 549856.0) + FrameCount * 5458u;
 
 float RandomValue()
 {
